# backend/docs/swagger/service.yaml
openapi: 3.0.0
info:
  title: DEU UCV API
  description: API del sistema de Extensión Universitaria de la UCV
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Ambiente Local
  - url: http://51.81.78.213:8081
    description: Servidor Remoto
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        count:
          type: integer
...
    User:
      type: object
      properties:
        id:
          type: string
        cedula:
          type: string
        email:
          type: string
        nombres:
          type: string
        apellidos:
          type: string
        biografia:
          type: string
        fecha_de_nacimiento:
          type: string
          description: DD-MM-YYYY
          pattern: '^\d{2}-\d{2}-\d{4}$'
          example: "14-07-1995"
        edad:
          type: integer
        genero:
          type: string
        nivel_educativo:
          type: string
        codigo_proveedor:
          type: string
        direccion:
          type: string
        creado_en:
          type: string
        foto_perfil:
          type: string
...
    Group:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        faculty:
          type: string
        objective:
          type: string
        location:
          type: string
        is_active:
          type: boolean
        owner:
          $ref: "#/components/schemas/User"
        created_at:
          type: string
        updated_at:
          type: string

    # ── Group Resource Requests ───────────────────────────────────────────────
    GroupResourceRequest:
      type: object
      properties:
        id:
          type: string
        grupo_id:
          type: string
        tipo:
          type: string
        contenido:
          type: string
        estado:
          $ref: "#/components/schemas/RequestStatus"
        creado_en:
          type: string
          format: date-time
        actualizado_en:
          type: string
          format: date-time
        razon_rechazo:
          type: string
          description: Only present when the request was rejected
    CreateGroupResourceRequestBody:
      type: object
      required: [ grupo_id, tipo ]
      properties:
        grupo_id:
          type: string
        tipo:
          type: string
        contenido:
          type: string

    # ── Request status ────────────────────────────────────────────────────────
    RequestStatus:
      type: string
      enum: [ created, under_review, approved, rejected, redirected ]

    # ── Course Requests ───────────────────────────────────────────────────────
...
    RejectWithReasonBody:
      type: object
      required: [ razon ]
      properties:
        razon:
          type: string
          description: Rejection reason. Saved and included in the email sent to the group's contact address. Blank values are rejected.
          example: Documentación incompleta
...
paths:
...
  /groups:
    get:
      summary: List groups
      tags: [ Groups ]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated group list
  /groups/{id}:
    get:
      summary: Get group by ID
      tags: [ Groups ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "404":
          description: Group not found
  /groups/requests:
    post:
      summary: Submit a new group registration request
      tags: [ Groups ]
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Group and request created
  /groups/requests/{id}:
    put:
      summary: Update group
      tags: [ Groups ]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated
        "404":
          description: Group not found
    delete:
      summary: Delete group
      tags: [ Groups ]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
        "404":
          description: Group not found

  # ── Group Resource Requests ───────────────────────────────────────────────
  /group-resource-requests:
    post:
      summary: Submit a group resource request
      tags: [ Group Resource Requests ]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupResourceRequestBody"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        "400":
          description: Validation error
        "401":
          description: Unauthorized

  # ── Admin — Group Resource Requests ───────────────────────────────────────
  /admin/group-resource-requests:
    get:
      summary: List group resource requests by faculty
      tags: [ Admin - Group Resource Requests ]
      security:
        - BearerAuth: []
      parameters:
        - name: faculty
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Paginated list of group resource requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  solicitudes:
                    type: array
                    items:
                      $ref: "#/components/schemas/GroupResourceRequest"
                  paginas:
                    $ref: "#/components/schemas/Pagination"
        "400":
          description: Invalid faculty
        "401":
          description: Unauthorized
  /admin/group-resource-requests/{id}:
    get:
      summary: Get group resource request by ID
      tags: [ Admin - Group Resource Requests ]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group resource request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResourceRequest"
        "404":
          description: Request not found
  /admin/group-resource-requests/{id}/approve:
    post:
      summary: Approve a group resource request
      tags: [ Admin - Group Resource Requests ]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Approved
        "404":
          description: Request not found (or not in reviewable state)
  /admin/group-resource-requests/{id}/reject:
    post:
      summary: Reject a group resource request
      tags: [ Admin - Group Resource Requests ]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectWithReasonBody"
      responses:
        "202":
          description: Rejected; the group's contact email is notified with the reason
        "400":
          description: razon missing or blank
        "404":
          description: Request not found (or not in reviewable state)

  # ── Course Cycle Close Requests (protected) ───────────────────────────────
...