# backend/internal/group_resource_requests/endpoints.go
package group_resource_requests

import (
	"context"
	"errors"
	"fmt"
	"net/http"
	"strings"

	"github.com/eaguilar88/deu/internal/entities"
	"github.com/eaguilar88/deu/internal/facultyscope"
	"github.com/eaguilar88/deu/internal/httperrors"
	"github.com/labstack/echo/v4"
	"go.uber.org/zap"
)

type Service interface {
	CreateGroupResourceRequest(ctx context.Context, req entities.GroupResourceRequest) (int64, error)
	ApproveGroupResourceRequest(ctx context.Context, reqID string) error
	RejectGroupResourceRequest(ctx context.Context, reqID, reason string) error
	GetGroupResourceRequestsByFaculty(ctx context.Context, faculty entities.Faculty, status string, pageScope entities.PageScope) ([]entities.GroupResourceRequest, int, entities.PageScope, error)
	GetGroupResourceRequestsByGroupID(ctx context.Context, groupID string, status string, pageScope entities.PageScope) ([]entities.GroupResourceRequest, entities.PageScope, error)
	GetPendingGroupResourceRequestsCountByFaculty(ctx context.Context, faculty entities.Faculty) ([]FacultyPendingCount, error)
	GetGroupResourceRequestByID(ctx context.Context, reqID string) (entities.GroupResourceRequest, error)
}

type Handler struct {
	svc Service
	log *zap.Logger
}

func NewHandler(svc Service, log *zap.Logger) *Handler {
	return &Handler{svc: svc, log: log}
}

func (h *Handler) RegisterGroupResourceRequestEndpoints(g *echo.Group) {
	gr := g.Group("/group-resource-requests")
	gr.POST("", h.CreateGroupResourceRequest)
	gr.GET("/group/:groupID", h.GetGroupResourceRequestsByGroupID)
}

func (h *Handler) RegisterGroupResourceRequestAdminEndpoints(g *echo.Group) {
	gr := g.Group("/group-resource-requests")
	gr.GET("", h.GetGroupResourceRequestsByFaculty)
	gr.GET("/pending-counts", h.GetPendingGroupResourceRequestsCountByFaculty)
	gr.GET("/:id", h.GetGroupResourceRequestByID)
	gr.POST("/:id/approve", h.ApproveGroupResourceRequest)
	gr.POST("/:id/reject", h.RejectGroupResourceRequest)
}

type createGroupResourceRequestInput struct {
	GroupID string `json:"grupo_id"  validate:"required,min=1"`
	Type    string `json:"tipo"      validate:"required,min=1"`
	Content string `json:"contenido"`
}

func (h *Handler) CreateGroupResourceRequest(c echo.Context) error {
	ctx := c.Request().Context()
	var input createGroupResourceRequestInput
	if err := c.Bind(&input); err != nil {
		return httperrors.NewBadRequest("invalid request body")
	}
	if err := c.Validate(input); err != nil {
		return httperrors.NewBadRequest("validation failed")
	}

	id, err := h.svc.CreateGroupResourceRequest(ctx, entities.GroupResourceRequest{
		GroupID: input.GroupID,
		Type:    input.Type,
		Content: input.Content,
		Status:  "under_review",
	})
	if err != nil {
		return httperrors.NewInternal(err)
	}

	return c.JSON(http.StatusCreated, CreateGroupResourceRequestResponse{
		ID: fmt.Sprintf("%d", id),
	})
}

func (h *Handler) ApproveGroupResourceRequest(c echo.Context) error {
	ctx := c.Request().Context()
	reqID := c.Param("id")
	if reqID == "" {
		return httperrors.NewBadRequest("request ID is required")
	}
	if err := h.svc.ApproveGroupResourceRequest(ctx, reqID); err != nil {
		if errors.Is(err, ErrNotFound) {
			return httperrors.NewNotFound("group resource request not found")
		}
		return httperrors.NewInternal(err)
	}
	return c.NoContent(http.StatusAccepted)
}

func (h *Handler) RejectGroupResourceRequest(c echo.Context) error {
	ctx := c.Request().Context()
	reqID := c.Param("id")
	if reqID == "" {
		return httperrors.NewBadRequest("request ID is required")
	}
	var req RejectGroupResourceRequestRequest
	if err := c.Bind(&req); err != nil {
		return httperrors.NewBadRequest("invalid request body")
	}
	req.Reason = strings.TrimSpace(req.Reason)
	if req.Reason == "" {
		return httperrors.NewBadRequest("razon es requerida")
	}
	if err := h.svc.RejectGroupResourceRequest(ctx, reqID, req.Reason); err != nil {
		if errors.Is(err, ErrNotFound) {
			return httperrors.NewNotFound("group resource request not found")
		}
		return httperrors.NewInternal(err)
	}
	return c.NoContent(http.StatusAccepted)
}

func (h *Handler) GetGroupResourceRequestsByFaculty(c echo.Context) error {
	ctx := c.Request().Context()
	faculty, err := facultyscope.Resolve(c, c.QueryParam("faculty"))
	if err != nil {
		return httperrors.NewBadRequest("invalid faculty")
	}

	status := c.QueryParam("status")

	var pageScope entities.PageScope
	//nolint:errcheck
	pageScope.GetPageFromVars(c.QueryParam("page"))
	//nolint:errcheck
	pageScope.GetPerPageFromVars(c.QueryParam("pageSize"))

	reqs, pendingCount, resultScope, err := h.svc.GetGroupResourceRequestsByFaculty(ctx, faculty, status, pageScope)
	if err != nil {
		return httperrors.NewInternal(err)
	}

	response := GetGroupResourceRequestsByFacultyResponse{
		Requests:     make([]GetGroupResourceRequestResponse, 0, len(reqs)),
		PendingCount: pendingCount,
		Pages:        resultScope,
	}
	for _, req := range reqs {
		response.Requests = append(response.Requests, toResponse(req))
	}
	return c.JSON(http.StatusOK, response)
}

func (h *Handler) GetGroupResourceRequestsByGroupID(c echo.Context) error {
	ctx := c.Request().Context()
	groupID := c.Param("groupID")
	if groupID == "" {
		return httperrors.NewBadRequest("groupID is required")
	}

	status := c.QueryParam("status")

	var pageScope entities.PageScope
	//nolint:errcheck
	pageScope.GetPageFromVars(c.QueryParam("page"))
	//nolint:errcheck
	pageScope.GetPerPageFromVars(c.QueryParam("pageSize"))

	reqs, resultScope, err := h.svc.GetGroupResourceRequestsByGroupID(ctx, groupID, status, pageScope)
	if err != nil {
		return httperrors.NewInternal(err)
	}

	response := GetGroupResourceRequestsResponse{
		Requests: make([]GetGroupResourceRequestResponse, 0, len(reqs)),
		Pages:    resultScope,
	}
	for _, req := range reqs {
		response.Requests = append(response.Requests, toResponse(req))
	}
	return c.JSON(http.StatusOK, response)
}

func (h *Handler) GetPendingGroupResourceRequestsCountByFaculty(c echo.Context) error {
	ctx := c.Request().Context()
	faculty, err := facultyscope.ResolveOptional(c, c.QueryParam("faculty"))
	if err != nil {
		return httperrors.NewBadRequest("invalid faculty")
	}
	counts, err := h.svc.GetPendingGroupResourceRequestsCountByFaculty(ctx, faculty)
	if err != nil {
		return httperrors.NewInternal(err)
	}
	return c.JSON(http.StatusOK, counts)
}

func (h *Handler) GetGroupResourceRequestByID(c echo.Context) error {
	ctx := c.Request().Context()
	reqID := c.Param("id")
	if reqID == "" {
		return httperrors.NewBadRequest("request ID is required")
	}
	req, err := h.svc.GetGroupResourceRequestByID(ctx, reqID)
	if err != nil {
		if errors.Is(err, ErrNotFound) {
			return httperrors.NewNotFound("group resource request not found")
		}
		return httperrors.NewInternal(err)
	}
	return c.JSON(http.StatusOK, toResponse(req))
}

func toResponse(r entities.GroupResourceRequest) GetGroupResourceRequestResponse {
	return GetGroupResourceRequestResponse{
		ID:              r.ID,
		GroupID:         r.GroupID,
		GroupName:       r.GroupName,
		Type:            r.Type,
		Content:         r.Content,
		Status:          r.Status,
		CreatedAt:       r.CreatedAt,
		UpdatedAt:       r.UpdatedAt,
		RejectionReason: r.RejectionReason,
	}
}